+++ types/tlds.d.ts
@@ -1,0 +1,5 @@
+declare module 'tlds' {
+  const tlds: Array<string>;
+
+  export default tlds;
+}